Compile eBPF programlets with -fno-stack-protector
The eBPF programlets cannot be built with -fno-stack-protector, as they
won't be linked to the supporting code, as it doesn't make sense in
their context. Add -fno-stack-protector to the build_ebpf line so that
they won't be built with this feature when the feature is enabled for
the userland parts of the project.